entity Curso
