+++ universitario/universitario/src/main/java/com/sistema/universitario/advices/CursoAdvice.java
@@ -1,0 +1,57 @@
+package com.sistema.universitario.advices;
+
+import com.sistema.universitario.exceptions.CursoNaoEncontradoException;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
+import org.springframework.core.convert.ConversionFailedException;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.context.request.WebRequest;
+import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Slf4j
+@Order(Ordered.HIGHEST_PRECEDENCE)
+@ControllerAdvice
+public class CursoAdvice extends ResponseEntityExceptionHandler{
+
+
+    @ExceptionHandler
+    public ResponseEntity NotFoundCurso(CursoNaoEncontradoException e){
+        ResponseEntity response = new ResponseEntity(e.getMessage(), HttpStatus.NOT_FOUND);
+        return response;
+    }
+
+   @ExceptionHandler
+    public ResponseEntity handleValidationExceptions(MethodArgumentNotValidException ex){
+        Map<String, String> errors = new HashMap<>();
+        ex.getAllErrors().forEach((error) -> {
+            String fieldName = ((FieldError) error).getField();
+            String errorMessage = ((FieldError) error).getDefaultMessage();
+            errors.put(fieldName, errorMessage);
+        });
+        log.error("error message {}", errors);
+        return new ResponseEntity(errors, HttpStatus.BAD_REQUEST);
+    }
+
+  /*  @ExceptionHandler(HttpMessageNotReadableException.class)
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ResponseEntity exceptionHandling(HttpMessageNotReadableException exception) {
+        Map<String, String> errors = new HashMap<>();
+        exception.getHttpInputMessage()..forEach((error) -> {
+            String fieldName = ((FieldError) error).getField();
+            String errorMessage = ((FieldError) error).getDefaultMessage();
+            errors.put(fieldName, errorMessage);
+        });
+        return new ResponseEntity(exception.getMessage(), HttpStatus.BAD_REQUEST);
+    }*/
+}
