+++ universitario/universitario/src/main/java/com/sistema/universitario/advices/DisciplinaAdvice.java
@@ -1,4 +1,4 @@
-package com.sistema.universitario.advice;
+package com.sistema.universitario.advices;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.ControllerAdvice;
