--- universitario/universitario/src/main/java/com/sistema/universitario/controller/CursoAdvice.java
@@ -1,40 +1,0 @@
-package com.sistema.universitario.controller;
-
-import com.sistema.universitario.exceptions.CursoNaoEncontradoException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-import java.util.HashMap;
-import java.util.Map;
-
-@Slf4j
-@ControllerAdvice
-public class CursoAdvice {
-
-
-    @ExceptionHandler
-    public ResponseEntity NotFoundCurso(CursoNaoEncontradoException e){
-        ResponseEntity response = new ResponseEntity(e.getMessage(), HttpStatus.NOT_FOUND);
-        return response;
-    }
-
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public Map<String, String> handleValidationExceptions(MethodArgumentNotValidException ex){
-        Map<String, String> errors = new HashMap<>();
-        ex.getAllErrors().forEach((error) -> {
-            String fieldName = ((FieldError) error).getField();
-            String errorMessage = ((FieldError) error).getDefaultMessage();
-            errors.put(fieldName, errorMessage);
-        });
-        log.error("error message {}", errors);
-        return errors;
-    }
-}
